+++ CHANGES.rst
@@ -5,9 +5,7 @@ Changelog
 2.0.4 (unreleased)
 ------------------
 
-- Add WeightedOrderedViewletManager which sort on viewlet weight.
-  It makes it easy to change order of viewlets by changing in sources.
-  [toutpt]
+- Nothing changed yet.
 
 
 2.0.3 (2012-12-15)